repo permissions: remove or update repository tests
