+++ test/unit/ta_test.rb
@@ -36,63 +36,4 @@ class TATest < ActiveSupport::TestCase
       assert @ta.is_assigned_to_grouping?(@grouping.id)
     end
   end
-
-  context 'If repo admin' do
-
-    setup do
-      @repo = Repository.get_class
-      MarkusConfigurator.stubs(:markus_config_repository_admin?).returns(true)
-    end
-
-    should 'grant repository_permissions when TA is added' do
-      ta = Ta.new
-      ta.user_name = 'just_another_admin'
-      ta.last_name = 'doe'
-      ta.first_name = 'john'
-
-      repo_names = Group.all.collect do |group|
-                     File.join(markus_config_repository_storage, group.repository_name)
-                   end
-      @repo.expects(:set_bulk_permissions).times(1).with(repo_names, {ta.user_name => Repository::Permission::READ_WRITE})
-      assert = ta.save
-    end
-
-    should 'revoke repository permissions when destroying a TA object' do
-      ta = Ta.make
-      repo_names = Group.all.collect do |group| File.join(markus_config_repository_storage, group.repository_name) end
-      @repo.expects(:delete_bulk_permissions).times(1).with(repo_names, [ta.user_name])
-      ta.destroy
-    end
-
-  end # end context
-
-  context 'If not repository admin' do
-
-    setup do
-      # set repository_admin false (NOT SURE HOW TO DO THIS NOW)
-      @repo = Repository.get_class
-      MarkusConfigurator.stubs(:markus_config_repository_admin?).returns(false)
-    end
-
-    teardown do
-      destroy_repos
-    end
-
-    should 'not remove repository permissions when deleting a TA' do
-      ta = Ta.make
-      @repo.expects(:delete_bulk_permissions).never
-      ta.destroy
-    end
-
-    should 'not grant repository permissions for newly created TAs' do
-      ta = Ta.new
-      ta.user_name = 'yet_another_admin'
-      ta.last_name = 'doe'
-      ta.first_name = 'john'
-
-      @repo.expects(:set_bulk_permissions).never
-      assert = ta.save
-    end
-  end
-
 end
